requirements.txt
numpy>=1.20.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.0
pydantic>=1.8.0,<2.0.0
scikit-learn>=1.0.0
pytest>=6.0.0
black>=21.5b2
isort>=5.8.0
mypy>=0.812
